requirements.txt
streamlit
selenium
selenium-stealth
beautifulsoup4
python-dotenv
requests
Pillow
webdriver_manager
python-multipart
pandas
langchain
langchain-community
langchain-core
ollama
openpyxl
undetected-chromedriver
